Origin-keyed agent clusters: update WPT for the rename
942aca884980a4dd6c582e3ca1a021769fd18f36 performed the minimal updates
to the implementation and tests to shift from "origin isolation" to
"origin-keyed agent clusters". This is part of the followup work to
fully reflect the rename, in this case focused on the web platform
tests.

The majority of changes are comment and <title> changes, although some
assertion messages are also updated.

Note that instead of renaming the top-level directory, we move it under
html/, since the the specification is now part of the HTML Standard.

Bug: 1158853
Change-Id: I008bb0b5f8c1f943849dfa5fd24cfc99a552f4dd